se arreglaron los porcentajes, se agrego validaciones a la creacion de clientes incluyendo la validacion de cuit

+++ EQ/EQNewOrderViewModel.h
@@ -47,8 +47,8 @@
 - (void)addItemQuantity:(int)quantity;
 - (NSNumber *)itemsQuantity;
 - (NSNumber *)subTotal;
-- (int)discountPercentage;
-- (int)discountValue;
+- (float)discountPercentage;
+- (float)discountValue;
 - (float)total;
 - (NSArray *)items;
 - (int)orderStatusIndex;
